backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库相关
sqlalchemy==2.0.25
aiosqlite==0.19.0

# JWT 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 数据验证
pydantic[email]==2.5.3
email-validator==2.1.0

# CORS 和静态文件
python-multipart==0.0.6

# CDN 文档
fastapi-cdn-host==0.8.0

# 脚本依赖
requests==2.31.0

bleach==6.3.0
fastapi==0.135.3
fastapi_cdn_host==0.10.0
pydantic==2.13.0
python_jose==3.5.0
SQLAlchemy==2.0.49
starlette==1.0.0

# 读取环境变量
python-dotenv==1.0.0

